Move Lux StatsTracking into a lux/stats module
